chore: continue cleanup
